+++ walkers/src/memory.rs
@@ -1,0 +1,54 @@
+use crate::{center::Center, position::AdjustedPosition, zoom::Zoom, InvalidZoom, Position};
+
+/// State of the map widget which must persist between frames.
+#[derive(Debug, Default, Clone)]
+#[cfg_attr(feature = "serde", derive(::serde::Serialize, ::serde::Deserialize))]
+pub struct MapMemory {
+    pub(crate) center_mode: Center,
+    pub(crate) zoom: Zoom,
+}
+
+impl MapMemory {
+    /// Try to zoom in, returning `Err(InvalidZoom)` if already at maximum.
+    pub fn zoom_in(&mut self) -> Result<(), InvalidZoom> {
+        self.center_mode = self.center_mode.clone().zero_offset(self.zoom.into());
+        self.zoom.zoom_in()
+    }
+
+    /// Try to zoom out, returning `Err(InvalidZoom)` if already at minimum.
+    pub fn zoom_out(&mut self) -> Result<(), InvalidZoom> {
+        self.center_mode = self.center_mode.clone().zero_offset(self.zoom.into());
+        self.zoom.zoom_out()
+    }
+
+    /// Set exact zoom level
+    pub fn set_zoom(&mut self, zoom: f64) -> Result<(), InvalidZoom> {
+        self.center_mode = self.center_mode.clone().zero_offset(self.zoom.into());
+        self.zoom = Zoom::try_from(zoom)?;
+        Ok(())
+    }
+
+    /// Returns the current zoom level
+    pub fn zoom(&self) -> f64 {
+        self.zoom.into()
+    }
+
+    /// Returns exact position if map is detached (i.e. not following `my_position`),
+    /// `None` otherwise.
+    pub fn detached(&self) -> Option<Position> {
+        self.center_mode.detached(self.zoom.into())
+    }
+
+    /// Center exactly at the given position.
+    pub fn center_at(&mut self, position: Position) {
+        self.center_mode = Center::Exact(AdjustedPosition {
+            position,
+            offset: Default::default(),
+        });
+    }
+
+    /// Follow `my_position`.
+    pub fn follow_my_position(&mut self) {
+        self.center_mode = Center::MyPosition;
+    }
+}
